인덱스 버퍼의 데이터 타입을 `Uint16Array`에서 `Uint32Array`로 변경하여 대규모 지형 데이터 처리 지원
**V4.4.0-Alpha-rnd 브런치.**

+++ src/display/terrain/core/geometry/TerrainGeometry.ts
@@ -17,7 +17,7 @@ class TerrainGeometry extends Geometry {
         const totalIndices = mainIndicesCount + skirtIndicesCount;
 
         const interleaveData = new Float32Array(totalVertexFloats);
-        const indexData = new Uint16Array(totalIndices);
+        const indexData = new Uint32Array(totalIndices);
 
         let vOffset = 0;
         let iOffset = 0;
